refactor: my-info API 경로 네이밍을 kebab-case로 통일
- shared/apis/myInfo를 shared/apis/my-info로 이동하고 관련 import를 일괄 갱신
- developer-detail, my-info, report 페이지의 참조 경로를 새 API 경로로 정리
- 파일 구조 컨벤션 문서의 my-info 표기를 현재 구조에 맞게 동기화

Made-with: Cursor
