+++ src/pages/my-info/my-info-setting.tsx
@@ -1,5 +1,5 @@
-import type { UpdateProfileRequest } from '@apis/myInfo/myInfo';
-import { myInfoQueries, updateMyProfile } from '@apis/myInfo/myInfo-queries';
+import type { UpdateProfileRequest } from '@apis/my-info/my-info';
+import { myInfoQueries, updateMyProfile } from '@apis/my-info/my-info-queries';
 import GithubIcon from '@assets/icons/github.svg?react';
 import GoogleIcon from '@assets/icons/google.svg?react';
 import { useUser } from '@clerk/clerk-react';
